Fixed internal component functions names / moved & adjusted tests
