+++ src/components/dashboard/FaceVerification/components/__tests__/CameraNotAllowedError.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { withThemeProvider } from '../../../../__tests__/__util__'
-import ImportedCameraNotAllowedError from '../components/CameraNotAllowedError'
+import { withThemeProvider } from '../../../../../__tests__/__util__'
+import ImportedCameraNotAllowedError from '../CameraNotAllowedError'
 
-const ErrorScreen = withThemeProvider(ImportedCameraNotAllowedError)
+const CameraNotAllowedError = withThemeProvider(ImportedCameraNotAllowedError)
 
 const screenState = {
   isValid: false,
@@ -17,12 +17,12 @@ const screenState = {
 
 describe('FaceVerification CameraNotAllowedErrorDialog', () => {
   it('renders without errors', () => {
-    const tree = renderer.create(<ErrorScreen screenProps={{ screenState }} />)
+    const tree = renderer.create(<CameraNotAllowedError screenProps={{ screenState }} />)
     expect(tree.toJSON()).toBeTruthy()
   })
 
   it('matches snapshot', () => {
-    const component = renderer.create(<ErrorScreen screenProps={{ screenState }} />)
+    const component = renderer.create(<CameraNotAllowedError screenProps={{ screenState }} />)
     const tree = component.toJSON()
     expect(tree).toMatchSnapshot()
   })
