+++ src/components/dashboard/FaceVerification/components/__tests__/DeviceOrientationError.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { withThemeProvider } from '../../../../__tests__/__util__'
-import ImportedDeviceOrientationError from '../components/DeviceOrientationError'
+import { withThemeProvider } from '../../../../../__tests__/__util__'
+import ImportedDeviceOrientationError from '../DeviceOrientationError'
 
-const ErrorScreen = withThemeProvider(ImportedDeviceOrientationError)
+const DeviceOrientationError = withThemeProvider(ImportedDeviceOrientationError)
 
 const screenState = {
   isValid: false,
@@ -17,12 +17,12 @@ const screenState = {
 
 describe('FaceVerification DeviceOrientationError', () => {
   it('renders without errors', () => {
-    const tree = renderer.create(<ErrorScreen screenProps={{ screenState }} />)
+    const tree = renderer.create(<DeviceOrientationError screenProps={{ screenState }} />)
     expect(tree.toJSON()).toBeTruthy()
   })
 
   it('matches snapshot', () => {
-    const component = renderer.create(<ErrorScreen screenProps={{ screenState }} />)
+    const component = renderer.create(<DeviceOrientationError screenProps={{ screenState }} />)
     const tree = component.toJSON()
     expect(tree).toMatchSnapshot()
   })
