+++ src/components/dashboard/FaceVerification/components/__tests__/DuplicateFoundError.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { withThemeProvider } from '../../../../__tests__/__util__'
-import ImportedDuplicateFoundError from '../components/DuplicateFoundError'
+import { withThemeProvider } from '../../../../../__tests__/__util__'
+import ImportedDuplicateFoundError from '../DuplicateFoundError'
 
-const ErrorScreen = withThemeProvider(ImportedDuplicateFoundError)
+const DuplicateFoundError = withThemeProvider(ImportedDuplicateFoundError)
 
 const screenState = {
   isValid: false,
@@ -17,12 +17,12 @@ const screenState = {
 
 describe('FaceVerification DuplicateFoundError', () => {
   it('renders without errors', () => {
-    const tree = renderer.create(<ErrorScreen screenProps={{ screenState }} />)
+    const tree = renderer.create(<DuplicateFoundError screenProps={{ screenState }} />)
     expect(tree.toJSON()).toBeTruthy()
   })
 
   it('matches snapshot', () => {
-    const component = renderer.create(<ErrorScreen screenProps={{ screenState }} />)
+    const component = renderer.create(<DuplicateFoundError screenProps={{ screenState }} />)
     const tree = component.toJSON()
     expect(tree).toMatchSnapshot()
   })
